Move fuel price service test suites to shared suite
