+++ pkg/services/fuelprice/diesel_fuel_price_storer_test.go
@@ -7,39 +7,11 @@ import (
 	"time"
 
 	"github.com/facebookgo/clock"
-	"github.com/stretchr/testify/suite"
 	"go.uber.org/zap"
 
 	"github.com/transcom/mymove/pkg/models"
-	"github.com/transcom/mymove/pkg/storage"
-	storageTest "github.com/transcom/mymove/pkg/storage/test"
 	"github.com/transcom/mymove/pkg/testdatagen"
-	"github.com/transcom/mymove/pkg/testingsuite"
 )
-
-type FuelPriceServiceSuite struct {
-	testingsuite.PopTestSuite
-	logger *zap.Logger
-	storer storage.FileStorer
-}
-
-func (suite *FuelPriceServiceSuite) SetupTest() {
-	suite.DB().TruncateAll()
-}
-
-func TestFuelPriceSuite(t *testing.T) {
-	// Use a no-op logger during testing
-	logger := zap.NewNop()
-
-	fakeS3 := storageTest.NewFakeS3Storage(true)
-
-	hs := &FuelPriceServiceSuite{
-		PopTestSuite: testingsuite.NewPopTestSuite(),
-		logger:       logger,
-		storer:       fakeS3,
-	}
-	suite.Run(t, hs)
-}
 
 func (suite *FuelPriceServiceSuite) TestStoreFuelPrices() {
 	testClock := clock.NewMock()
